ISAICP-4572: Create an utility trait to avoid RDF entity storage code dulplication.
